+++ packages/core/CHANGELOG.md
@@ -8,6 +8,12 @@ Versioning].
 
 ## [Unreleased]
 
+- _No changes yet_
+
+## [0.1.18] - 2021-04-17
+
+- Fix accidental mangling in HTML content.
+- Fix various cases where attributes in HTML would not be mangled.
 - Improve performance of the built-in CSS language plugin.
 - Improve performance of the built-in JavaScript language plugin.
 - Support mangling attributes selectors with unquoted values.